Moving PreTestDeclarations to common
