+++ src/modules/tests/__tests__/tests.reducer.spec.ts
@@ -1,6 +1,6 @@
 import { testsReducer } from '../tests.reducer';
 import * as candidateReducer from '../journal-data/common/candidate/candidate.reducer';
-import * as preTestDeclarationsReducer from '../pre-test-declarations/pre-test-declarations.reducer';
+import * as preTestDeclarationsReducer from '../pre-test-declarations/common/pre-test-declarations.reducer';
 import { PreTestDeclarations } from '@dvsa/mes-test-schema/categories/common';
 import { CatBUniqueTypes } from '@dvsa/mes-test-schema/categories/B';
 import { TestsModel } from '../tests.model';
