+++ src/modules/tests/pre-test-declarations/__tests__/pre-test-declarations.reducer.spec.ts
@@ -1,11 +1,11 @@
-import { preTestDeclarationsReducer, initialState } from '../pre-test-declarations.reducer';
+import { preTestDeclarationsReducer, initialState } from '../common/pre-test-declarations.reducer';
 import {
   ToggleInsuranceDeclaration,
   ToggleResidencyDeclaration,
   SignatureDataChanged,
   SignatureDataCleared,
   ClearPreTestDeclarations,
-} from '../pre-test-declarations.actions';
+} from '../common/pre-test-declarations.actions';
 import { PreTestDeclarations } from '@dvsa/mes-test-schema/categories/common';
 
 describe('PreTestDeclarations reducer', () => {
