--- src/modules/tests/pre-test-declarations/cat-a-mod2/__tests__/pre-test-declarations.selector.spec.ts
@@ -1,32 +1,0 @@
-import {
-  getInsuranceDeclarationStatus,
-  getResidencyDeclarationStatus,
-  getSignatureStatus,
-} from '../pre-test-declarations.cat-a-mod2.selector';
-import { PreTestDeclarations } from '@dvsa/mes-test-schema/categories/common';
-
-describe('PreTestDeclarations Cat A Mod2 selector', () => {
-  const state: PreTestDeclarations = {
-    insuranceDeclarationAccepted: true,
-    residencyDeclarationAccepted: false,
-    preTestSignature: 'sig',
-  };
-
-  describe('getInsuranceDeclarationStatus', () => {
-    it('should return the insurance declaration status', () => {
-      expect(getInsuranceDeclarationStatus(state)).toBe(true);
-    });
-  });
-
-  describe('getResidencyDeclarationStatus', () => {
-    it('should return the residency declaration status', () => {
-      expect(getResidencyDeclarationStatus(state)).toBe(false);
-    });
-  });
-
-  describe('getSignatureStatus', () => {
-    it('should return the signature status', () => {
-      expect(getSignatureStatus(state)).toEqual('sig');
-    });
-  });
-});
