--- src/modules/tests/pre-test-declarations/cat-a-mod2/pre-test-declarations.cat-a-mod2.selector.ts
@@ -1,5 +1,0 @@
-import { PreTestDeclarations } from '@dvsa/mes-test-schema/categories/common';
-
-export const getInsuranceDeclarationStatus = (decs: PreTestDeclarations) => decs.insuranceDeclarationAccepted;
-export const getResidencyDeclarationStatus = (decs: PreTestDeclarations) => decs.residencyDeclarationAccepted;
-export const getSignatureStatus = (decs: PreTestDeclarations) => decs.preTestSignature;
