+++ src/modules/tests/pre-test-declarations/common/pre-test-declarations.selector.ts
@@ -1,4 +1,4 @@
-import { PreTestDeclarations } from '@dvsa/mes-test-schema/categories/common';
+import { PreTestDeclarations } from './node_modules/@dvsa/mes-test-schema/categories/common';
 
 export const getInsuranceDeclarationStatus = (decs: PreTestDeclarations) => decs.insuranceDeclarationAccepted;
 export const getResidencyDeclarationStatus = (decs: PreTestDeclarations) => decs.residencyDeclarationAccepted;
