+++ src/pages/waiting-room/cat-be/waiting-room.cat-be.page.ts
@@ -6,14 +6,14 @@ import { Store, select } from '@ngrx/store';
 import { StoreModel } from '../../../shared/models/store.model';
 import * as waitingRoomActions from '../waiting-room.actions';
 import { Observable } from 'rxjs/Observable';
-import { getPreTestDeclarations } from '../../../modules/tests/pre-test-declarations/pre-test-declarations.reducer';
+import { getPreTestDeclarations } from '../../../modules/tests/pre-test-declarations/common/pre-test-declarations.reducer';
 import * as preTestDeclarationsActions
-  from '../../../modules/tests/pre-test-declarations/pre-test-declarations.actions';
+  from '../../../modules/tests/pre-test-declarations/common/pre-test-declarations.actions';
 import {
   getInsuranceDeclarationStatus,
   getResidencyDeclarationStatus,
   getSignatureStatus,
-} from '../../../modules/tests/pre-test-declarations/pre-test-declarations.selector';
+} from '../../../modules/tests/pre-test-declarations/common/pre-test-declarations.selector';
 import { getCandidate } from '../../../modules/tests/journal-data/cat-be/candidate/candidate.cat-be.reducer';
 import {
   getCandidateName, getCandidateDriverNumber, formatDriverNumber, getUntitledCandidateName,
